perf(metric-learning): update template lib generate and similarity match usage

+++ tools/match/generate_template_lib.py
@@ -4,21 +4,31 @@
 @date: 2022/1/18 下午3:33
 @file: generate_templte_lib.py
 @author: zj
-@description: 对训练数据进行采集，创建模板特征库，每个类别最多采集N条数据
+@description: 对训练数据进行采集，创建模板特征库
 """
 
 import os
 import pickle
+import argparse
 
 import numpy as np
 from tqdm import tqdm
+
+
+def parse_args():
+    parser = argparse.ArgumentParser()
+    parser.add_argument('csv_path', type=str, default=None, help='csv path')
+    parser.add_argument('pkl_path', type=str, default=None, help='pkl path')
+
+    args = parser.parse_args()
+    print(args)
+    return args
 
 
 def load_data(csv_path):
     assert os.path.isfile(csv_path), csv_path
 
     data_array = np.loadtxt(csv_path, delimiter=',', dtype=float)
-
     return data_array
 
 
@@ -39,9 +49,17 @@ def process(data_array):
 
 
 if __name__ == '__main__':
-    csv_path = './outputs/train_logits.csv'
+    args = parse_args()
+    csv_path = args.csv_path
+    pkl_path = args.pkl_path
+
+    if not os.path.exists(csv_path):
+        raise ValueError(f'{csv_path} does not exists')
+    if os.path.exists(pkl_path):
+        raise ValueError(f'{pkl_path} has existed')
+
     data_array = load_data(csv_path)
 
     match_dict = process(data_array)
-    with open('./outputs/train_template_lib.pkl', 'wb') as f:
+    with open(pkl_path, 'wb') as f:
         pickle.dump(match_dict, f)
